feat(sdk)!: Retire redundant suffixes and per-host symbols
File suffixes, continuing the target-framework vocabulary:

  *.Apple.cs     -> *.UIKit.cs, whose rule was byte-for-byte identical
                    (IsApple and IsAppleUIKit are both "IsIOS or IsTvOS").
                    The 29 files are renamed; SoftwareBitmap had one of each
                    and its two halves are merged.
  *.reference.cs -> dropped from the SDK. It selects on UnoRuntimeIdentifier,
                    which nothing outside Uno.UWP / Uno.Foundation /
                    Uno.UI.Dispatching sets, so for an application it compiled
                    nowhere and warned nowhere. Those three projects keep it
                    through the repository's own targets.
  *.iOSmacOS.cs  -> dropped, it named the native macOS target removed in 7.0.
  *.Xamarin.cs   -> dropped from the repository targets.

Preprocessor symbols, closing unoplatform/uno#17684: the eight
HAS_UNO_SKIA_<host> symbols and their __UNO_SKIA_<host>__ counterparts are
removed. They read as a compile-time host discriminator but cannot be one --
the SDK references every desktop host package together, so a netX.0-desktop
head defined all of them at once. Nothing branched on them. HAS_UNO_SKIA and
__UNO_SKIA__ are untouched; the XAML generator emits the former.
