+++ doc/articles/uno-development/uno-internals-macos.md
@@ -7,7 +7,3 @@ uid: Uno.Contributing.macOS
 macOS is supported through the Skia rendering backend. The native macOS target (where `UIElement` inherited from `NSView`) has been removed. macOS now uses the same Skia-based rendering pipeline as other desktop platforms (Windows and Linux).
 
 For details on how the Skia backend works, see the [overview article](uno-internals-overview.md).
-
-## The `.iOSmacOS.cs` suffix
-
-Some files in the codebase use the `*.iOSmacOS.cs` suffix. This suffix predates the removal of the native macOS target and is now simply Apple UIKit code — these files are compiled when targeting `net10.0-ios`.
