+++ src/Uno.UI.Runtime.Skia.Headless/buildTransitive/Uno.WinUI.Runtime.Skia.Headless.targets
@@ -7,7 +7,7 @@
 	<Target Name="_UnoSkiaHeadlessFeatureDefines" AfterTargets="PrepareForBuild">
 
 		<PropertyGroup>
-			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;HAS_UNO_SKIA_HEADLESS;__DESKTOP__;__UNO_SKIA__;__UNO_SKIA_HEADLESS__</DefineConstants>
+			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;__DESKTOP__;__UNO_SKIA__</DefineConstants>
 		</PropertyGroup>
 
 	</Target>
