+++ src/Uno.UI.Runtime.Skia.Linux.FrameBuffer/buildTransitive/Uno.WinUI.Runtime.Skia.Linux.FrameBuffer.targets
@@ -7,7 +7,7 @@
 	<Target Name="_UnoSkiaLinuxFBFeatureDefines" AfterTargets="PrepareForBuild">
 
 		<PropertyGroup>
-			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;HAS_UNO_SKIA_LINUX_FB;__DESKTOP__;__UNO_SKIA__;__UNO_SKIA_LINUX_FB__</DefineConstants>
+			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;__DESKTOP__;__UNO_SKIA__</DefineConstants>
 		</PropertyGroup>
 
 	</Target>
