+++ src/Uno.UI.Runtime.Skia.MacOS/buildTransitive/Uno.WinUI.Runtime.Skia.MacOS.targets
@@ -7,7 +7,7 @@
 	<Target Name="_UnoSkiaMacOSFeatureDefines" AfterTargets="PrepareForBuild">
 
 		<PropertyGroup>
-			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;HAS_UNO_SKIA_MACOS;__DESKTOP__;__UNO_SKIA__;__UNO_SKIA_MACOS__</DefineConstants>
+			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;__DESKTOP__;__UNO_SKIA__</DefineConstants>
 		</PropertyGroup>
 
 	</Target>
