+++ src/Uno.UI.Runtime.Skia.Win32/buildTransitive/Uno.WinUI.Runtime.Skia.Win32.targets
@@ -7,7 +7,7 @@
 	<Target Name="_UnoSkiaWin32FeatureDefines" AfterTargets="PrepareForBuild">
 
 		<PropertyGroup>
-			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;HAS_UNO_SKIA_WIN32;__DESKTOP__;__UNO_SKIA__;__UNO_SKIA_WIN32__</DefineConstants>
+			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;__DESKTOP__;__UNO_SKIA__</DefineConstants>
 		</PropertyGroup>
 
 	</Target>
