+++ src/Uno.UI.Runtime.Skia.X11/buildTransitive/Uno.WinUI.Runtime.Skia.X11.targets
@@ -7,7 +7,7 @@
 	<Target Name="_UnoSkiaX11FeatureDefines" AfterTargets="PrepareForBuild">
 
 		<PropertyGroup>
-			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;HAS_UNO_SKIA_X11;__DESKTOP__;__UNO_SKIA__;__UNO_SKIA_X11__</DefineConstants>
+			<DefineConstants>$(DefineConstants);UNO_REFERENCE_API;HAS_UNO_SKIA;__DESKTOP__;__UNO_SKIA__</DefineConstants>
 		</PropertyGroup>
 
 	</Target>
