+++ src/Uno.WinRT/ApplicationModel/Package.UIKit.cs
