+++ src/Uno.WinRT/ApplicationModel/PackageId.UIKit.cs
