+++ src/Uno.WinRT/Extensions/NSUrlExtensions.UIKit.cs
