+++ src/Uno.WinRT/Gaming/Input/Gamepad.UIKit.cs
