+++ src/Uno.WinRT/Media/Playback/Internal/VideoSurface.UIKit.cs
