+++ src/Uno.WinRT/Media/Playback/MediaPlayer.UIKit.cs
