+++ src/Uno.WinRT/Networking/Connectivity/ConnectionProfile.UIKit.cs
