+++ src/Uno.WinRT/Networking/Connectivity/Internal/NetworkReachability.UIKit.cs
