+++ src/Uno.WinRT/Networking/Connectivity/NetworkInformation.UIKit.cs
