+++ src/Uno.WinRT/Storage/ApplicationDataContainer.UIKit.cs
