+++ src/Uno.WinRT/Storage/Helpers/StorageFileHelper.UIKit.cs
