+++ src/Uno.WinRT/Storage/Internal/NSUrlExtensions.UIKit.cs
