+++ src/Uno.WinRT/Storage/Internal/SecurityScopeManager.UIKit.cs
