+++ src/Uno.WinRT/Storage/Pickers/Internal/UTTypeMapper.UIKit.cs
