+++ src/Uno.WinRT/Storage/StorageFile.UIKit.cs
