+++ src/Uno.WinRT/System/Launcher.UIKit.cs
